Fix some small bugs in WVector

+++ src/FileContainer/FileReader.h
@@ -23,7 +23,7 @@ public:
 
 inline auto FileReader::read(void *to, size_t size) const -> int
 {
-    if (_delta + size > _data->size())
+    if (_delta + size > static_cast<unsigned long>(_data->size()))
         return -1;
 
     WCommonScript::WMemcpy(to, _data->constData() + _delta, size);
